# gateway-go/openapi.yaml
openapi: 3.1.0
info:
  title: Defence Cyber Portal - Integration Gateway API
  description: High-throughput ingestion gateway for security alerts, webhooks, syslog, and polling connectors.
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1
    description: Local development gateway
paths:
  /health:
    get:
      summary: Health check and connector operational metrics
      responses:
        '200':
          description: Gateway is healthy and connected to event bus
  /events:
    post:
      summary: Submit security event to canonical ingestion pipeline
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
            example: Bearer soc-service-token-xyz
        - in: header
          name: X-Tenant-ID
          required: true
          schema:
            type: string
            example: tenant_acme
        - in: header
          name: X-Source-ID
          required: true
          schema:
            type: string
            example: edr-prod-01
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
            example: edr-prod-01:alert-883921
        - in: header
          name: traceparent
          required: false
          schema:
            type: string
            example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Event accepted and published to event bus topic
        '200':
          description: Duplicate event identified; original event ID returned
        '400':
          description: Bad request, schema failure, or payload hash mismatch
        '401':
          description: Unauthorized
        '429':
          description: Rate limited with Retry-After header
        '503':
          description: Event bus temporarily unavailable
  /alerts:
    post:
      summary: Submit prioritized security alert
      responses:
        '202':
          description: Alert accepted
  /webhooks/{connector}:
    post:
      summary: Ingest vendor push alert with HMAC signature verification
      parameters:
        - in: path
          name: connector
          required: true
          schema:
            type: string
        - in: header
          name: X-Signature
          required: false
          schema:
            type: string
      responses:
        '202':
          description: Webhook accepted